Move data manipulation to database
This is better as it ensures that if we need to get some extra data from
the database, the database will not be out of sync with us. Now the python
server will only serve as a bridge to MQTT and will receive full information
on what to publish from the database.